Add gems, correct tests, reformat helpers

+++ app/helpers/followings_helper.rb
@@ -6,11 +6,13 @@ module FollowingsHelper
     following = Following.find_by(user_id: current_user.id, follow_id: n_user.id) if following.nil?
 
     if current_user == n_user
-      link_to 'Unfollow', following_path(id: following.id), method: :delete, class: 'btn btn-danger'
+      content_tag :button, (link_to 'Unfollow', following_path(id: following.id), method: :delete, data: { confirm: 'Are you sure?' }), class: 'btn btn-danger', type: 'button'
     elsif current_user.follow?(n_user)
-      link_to 'Unfollow', following_path(id: following.id), method: :delete, class: 'btn btn-danger'
+      content_tag :button, (link_to 'Unfollow', following_path(id: following.id), method: :delete, data: { confirm: 'Are you sure?' }), class: 'btn btn-danger', type: 'button'
+      #link_to 'Unfollow', following_path(id: following.id), method: :delete, class: 'btn btn-danger'
     else
-      link_to 'Follow', followings_path(user_id: current_user.id, follow_id: n_user.id), method: :post, class: 'btn btn-primary'
+      content_tag :button, (link_to 'Follow', followings_path(user_id: current_user.id, follow_id: n_user.id), method: :post), class: 'btn btn-succes', type: 'button'
+      #link_to 'Follow', followings_path(user_id: current_user.id, follow_id: n_user.id), method: :post, class: 'btn btn-primary'
     end
   end
